Input Manager Added & Player One Controls Binded
I added in the input manager and binded the following controls to their respective buttons:
-Move(Left Stick)
-Rotate(Right Stick)
-Attack(A)
-Special Attack(B & Y)
-Interact(X)
-Options(Menu)
Move is the only one that has code to go with it as of this push. You can move the white square sprite.
I also added a PlayerOneBehvaiour script that is attached to the white square sprite.