Added Local Co-op
Ignore the last push, where I had added all the bindings and whatnot. I learned I couldn't make it co-op like that.

This code allows players to join by pressing any button on their controller, then they can move via the left joystick.

The players are set as a prefab right now because that's how the player input manager will take them. An empty game object has the player input manager on it, and it is just called 'Player Input Manager' in the hierarchy, that's what allows the players to actually join.